PM_agent_streamlit/requirements.txt
streamlit
langchain-core
langchain-openai
langchain-community
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
langgraph